+++ ethereum/eip712/encoding.go
@@ -154,9 +154,14 @@ func decodeProtobufSignDoc(signDocBytes []byte) (apitypes.TypedData, error) {
 		return apitypes.TypedData{}, err
 	}
 
-	// Until support for these fields is added, throw an error at their presence
-	if body.TimeoutHeight != 0 || len(body.ExtensionOptions) != 0 || len(body.NonCriticalExtensionOptions) != 0 {
-		return apitypes.TypedData{}, errors.New("body contains unsupported fields: TimeoutHeight, ExtensionOptions, or NonCriticalExtensionOptions")
+	// Until support for these fields is added, throw an error at their presence.
+	if body.TimeoutHeight != 0 ||
+		body.GetTimeoutTimestamp() != nil ||
+		len(body.ExtensionOptions) != 0 ||
+		len(body.NonCriticalExtensionOptions) != 0 {
+		return apitypes.TypedData{}, errors.New(
+			"body contains unsupported fields: TimeoutHeight, TimeoutTimestamp, ExtensionOptions, or NonCriticalExtensionOptions",
+		)
 	}
 
 	if len(authInfo.SignerInfos) != 1 {